requirements.txt
flask
transformers
torch
pillow
pillow-heif
gunicorn
